+++ src/bandersnatch/mirror.py
@@ -337,7 +337,6 @@ class BandersnatchMirror(Mirror):
         await loop.run_in_executor(
             self.storage_backend.executor, self.sync_simple_pages, package
         )
-        # XMLRPC PyPI Endpoint stores raw_name so we need to provide it
         await loop.run_in_executor(
             self.storage_backend.executor,
             self.record_finished_package,